Clean up lists, document formats.
The only lists we have now are SONGS, ALLSONGS, and ZONGS; only the
latter is derived from zongbook.tex; it works correctly (i.e. makes
empty lists and doesn't crash) even if zongbook.tex is missing.

Note that we are not yet setting style options from definitions on the
command line; that obviously has to be done before the PDFs we make
here will work properly in a looseleaf, duplex-printed songbook.