remove include/, {HEADER,README}.html
The stuff in include is old, and no longer useful -- both boilermaker.pl
and replace-template-file.pl are buggy and have been superceded by
mustache and Makefile templates.

HEADER.html and README.html are redundant given that we have README.md
and that the public version of MakeStuff is on GitHub.