starting page-to-template-data
This is a rather hacky python program that converts a web page
(typically index.html) into a body-contents fragment with an email-style
Subject: header derived from the title.  It is used for converting a
complete page into input for a templating engine (e.g. mustache).

RFC822-style headers are used in part because we're already using them
for blog posts (see MakeStuff/blogging), and in part because they don't
have to be quoted.  Note that there's no real point in making the header
into YAML frontmatter, because the body isn't the outermost template,
and more metadata is going to be added at build time.