add import-blog-entry
This imports MakeStuff blog entries into a Jekyll _posts directory.
Filenames are converted from yyyy/mm/dd--name to yyyy-mm-dd-name format,
and the front matter is put into YAML format no matter what format it
had originally.

The "Subject" header is converted to "title", tags get surrounded by
brackets to make a list, and tags of "curmudgeon" are removed from the
list.  (That last could be generalized, but that can wait.)