remove fixup*
These were one-shot fixes, and were written before I was fully
comfortable using find, sed, and so on.  These days I mostly just do
things like that on the command line and document them in one of my
ubiquitous to.do files, and although anything really tricky will still
be done with a script, if it's really one-time it's not worth keeping.