Build song PDFs by copying from Lyrics
This copies %/lyrics.pdf from ../Lyrics rather than building it in
place.  We are careful to leave a copy in Lyrics as well.  Comments in
songs.make mention a couple of options for creating the PDFs directly in
the song directories.  That would let us make them with a different
format, e.g. compact rather than looseleaf, and might still be a good
idea.