Improve page headers, document zingers
Headers now have page numbers on the outside where they're easy to
spot when you're flipping through a songbook.

Document the fact that in a Lyrics directory, a local zingers.sty can be
used to customize the page style -- in particular, the footers.  They
can be overridden again in zongbook.tex, so that broadsides and
songbooks can have different footers.