replace multiple "-"s with single ones
All it took was changing s/ +/-/ to s/[ -]+/-/ .
This change makes sure that we don't get multiple dashes when we
sluggify a title like "this -- that".  Since we use "--" to separate
the posting date from the slug, it makes certain that we don't get false
matches even if some titles contain dates.