Improve efficiency of several variable definitions
Make DOTDOT, HOST, and EXCLUDES recursive.  They're only used in the
rsync* targets, so most of the time they don't need to be expanded (and
run in the shell).

PLEASE NOTE:  any config.make etc. that overrides them will have to
change := to =; hopefully there won't be many of these.

BASEREL and TOOLREL are both more efficient and more readable if we use
`realpath --relative-to` rather than the for loops they were using.

BASEDIR still uses a for loop, but it doesn't bother testing for Tools
because all occurrences have (hopefully) been changed to point to
MakeStuff.