Fix BASEREL and TOOLREL definitions
Using `TOOLREL:=$(shell realpath --relative-to=. $(TOOLDIR))` breaks the
rules for %/lyrics.* in music/songs.make.  It's not entirely clear why,
but presumably it's because BASEREL ends up pointing to the _physical_
parent directory of MakeStuff.

It was still possible to simplify the loop by not looking for Tools,
which is unnecessary now that we're no longer using it.  The rule for
BASEREL was also simplified by having the TOOLREL shell script return
./MakeStuff instead of Makestuff, eliminating the need for a conditional
to handle the special case.