Add TeX/README.md, with details of FlkTex markup
TeX/README.md now contains a pretty complete description of the FlkTex
markup language and a discussion of the various songbook formats.

It still needs a man page or the equivalent for `flktran.pl`,
`index.pl`, and `sort-by-title`, and it would eventually be a good idea
to pull in `transpose`, `songinfo`, and `lyrics.make` from ../music.