Troubleshoot problems with non-Flk parts
The text inside of note, inset, etc., and after `\end{song}` is plain
LaTeX, not FlkTex.  Setting `plain++` handles most of this well, except
that there are a lot of LaTeX constructs, lists and math in particular,
that it _doesn't_ handle but ought to if we want to make the HTML look
decent.  I'm not sure there's any decent way to translate them into
ChordPro, either.

In many cases we could simply leave off after `\end{song}` and
hand-translate (or pandoc) the remainder if that's useful (e.g. in
heavily-annotated songs like QV and _The Stuff that Dreams are Made
Of_), but that should be optional (i.e. a command-line option).