update translations in songinfo to HTML5
Physical markup tags, e.g. `<b>` and `<tt>`, are deprecated in HTML5.
Replace them with `<strong>` and `<code>` respectively.  Note that this
is a potentially lossy translation -- there are other reasons why one
might want to italicize or boldface something -- but those ambiguities
are present in the original TeX markup, so we're not actually losing
all that much.