Move "done" items from to.do to Archive/2020.done
The next step will be to tag this commit to make it easy to find, then
delete Archive altogether.  The pseudo-changelist in to.do and Archive/*
is _extremely_ inaccurate -- there are multi-year gaps, for example.  A
more complete (if somewhat sketchy) changelist can be extracted from the
Git log.  The tag will make it possible to retrieve the old changelog
if, for some inexplicable reason, that becomes necessary.