Refactor how entry.make gets ENTRY, DRAFT
The idea is to make DRAFT and ENTRY recursively-expanded, so that they
will respond properly to target-specific definitions of name and title.
This allows us to use target-specific variables in mixins like, e.g.,
`make thanks`.

It doesn't quite work, because targets, dependencies with computed
names, and automatic variables are outside of recipes.  At this point, a
sub-make is still necessary if you need to override the definition of
$(name).  That will be fixed in the next commit.