Refactor computed targets out of entry.make
Now that we have fixed $(ENTRY) and $(DRAFT) so that they are expanded
recursively and depend only on $(name), the next step is to stop using
them as targets.  So instead of
```
entry: $(ENTRY)
$(ENTRY):
	xyz
```
we now have
```
entry:
	xyz
```

Validation targets like `name-required` are safe because if we define
target-specific variables, they get carried over to the dependencies
of that target.

The `thank:` mixin gives an example of how to use target-specific
variables.