Make target-specific variables work sensibly
There are now two sets of variables, for the two different contexts of
draft creation and posting.  The posting context variables depend only
on the variables defined on the command line, or a .draft symlink if
there is one.  That means that `make post` should work as before, and it
should work with .draft pointing at a draft file rather than an entry.

The variables for the posting context are: draft, entry, name, and
title, all lower case.  There is one breaking change: if you want to
pass a direct pointer to an entry on the command line, you will need to
use `entry=...` rather than `ENTRY=...` as before.  That makes the
command line UI more consistent; my guess is that nobody uses it
anyway.

The variables for the draft creation context are DRAFT, ENTRY, NAME, and
TITLE.  Unlike the lowercase posting variables, these can be re-assigned
as target-specific variables, for use in shortcut targets like `thanks`
and `done`.  They are *not* affected by the presence of a `.draft`
symlink; that led to very confusing behavior in earlier experiments.  In
addition, you never need to use override to make target-specific
definitions.  You do need to use `:=` assignment -- otherwise
definitions that modify them (say, by adding a prefix) would cause
recursion if Make didn't flag them as an error.

Of course, the old way of doing things by passing a new name, title,
etc. on the command line of a sub-make still works, and is arguably
simpler in most cases.  But not as maintainable, and not nearly as much
fun.  Whether it was worthwhile to spend three days of hacking to save a
few milliseconds over the next several years of blogging is a question I
prefer to leave unanswered.