Use ?= for default tag values in songs.
WEB_OK_TAGS and NOT_OK_TAGS were assigned using =, which meant that they
replaced the non-default values assigned in `.config.make`.  This showed
up when someone looked for lyrics for The October Country in
`steve.savitzky.net/Songs`.  Using `?=` in `songs.make` keeps this from
happening.

It wouldn't have happened at all if .config.make was included _after_
`songs.make`, or _in_ it, but it isn't.  Using `?=` handles both cases,
so we don't have to worry if we move that around.