Fix bugs in `make post`
Rather than storing the result of POST_CMD in a shell variable, we use a
recursively-expanded make variable.  In addition to being simpler, it
makes what the command returns visible in make's output.

Add a -v flag to charm-wrapper so that we can show what charm returns --
its error messages appear to come out on STDOUT.

Remove `@` prefix on the template-expansion echo commands, so that if
there's an extraneous space at the end of $(PFX) it's easy to spot.