Fix multiple bugs in pre-post, post, and posted
Now that we are allowing .draft to point to either a draft or an entry,
we need to remove pre-post's dependency on `draft-or-entry-required`.

In the recipd for post:, putting the url returned by $(POSTCMD) in a
recursively-expanded make variable rather than a shell variable would
have been a good idea except that it would get run even if make had been
run with the -n flag.

Some quoting issues had crept in to `post:` and `posted:` -- if there
are single quotes around the sed command, we only need one backslash in
`\n`, because escapes aren't recognized inside of single quotes.

This undoes some of the changes in made in commit
54ffc49dd46b2171384a424c31e9bb96f5adce9a.  The other changes there are
still good, which is why we don't simply revert it.