Fix post problem caused by charm output
As it turns out, the problem that caused the sed error:
```
    `\sed: -e expression #1, char 45: unterminated `s' command
```
was _not_ a quoting problem.  The problem was that in order to reveal
the error message returned by charm, we stopped redirecting its output
to /dev/null.  Trouble is, charm prints "Posted." on success.  That was
getting in front of the url returned by the `wget` (essentially
last-post) at the end.  The resulting space is what was messing up sed.

Now, we take _whatever_ charm-wrapper or any other posting command
returns, and examine it to see whether it's a URL. If it doesn't start
with `http`, `file`, `.`, or a digit, we invoke the `last-post` script.
It may, of course, return the wrong thing because it can't find anything
posted under a filter.  We'll fix that later.