make pull-all check for renamed master
If the master branch has been renamed to main on the origin, also rename
it on the clone that we're updating.

The way we do the check is clumsy and inefficient, and only works for
master->main; this is not ideal but it works for the only case it's
likely to run into over the next few months.