remove bogus tab after "posts" in -Y
It's not clear why I thought a tab would be useful, but it isn't, and it
messes up formatting if you right-shift the block.  That's on the
assumption that there won't be a month with more than 99 posts, but even
if there is the resulting offset won't matter much.

Worry about that if it happens.