Add blogging/save.make with save and Save targets
`blogging/save.make` is a MakeStuff plugin that defines two targets:
`save:` and `Save:`.  It is usually included along with `entry.make` in
a `.blog.defaults.make` or similar config file.  To make it apply to a
whole tree, include it in `$(BASEDIR)/.site/config.make`.

The `make save` target is meant for making "trivial" commits; its main
use is for saving `to.do` or a draft after adding a line or two; or to
save work before moving to a different computer.  A second `make save`
on the same host will use `commit --amend` and `push --force-with-lease`
to overwrite the commit made by the previous save.  If the current
commit does not start with `Saved on`, `make save` will do a `make push`
instead.

The `make Save` target is similar but will also overwrite a commit
created by `make push`.  Whether Save: -- and save:, for that matter --
should overwrite a commit made on a different host is an open question,
as is the question of whether to require the overwritten commit to have
been made on the same day.

There is a tweak variable, settable on the command line or in a config
file, called `SAVE_MAY_AMEND_PUSH`, which allows save: to clobber a
`Push from` commit.  TODO:  There should also be one to control whether
to check the host, and another to control comparing dates.  It's really
overkill given that there are less than a handful of people who use
MakeStuff for blogging.  Possibly only one.

CAUTION:  `make save` and `make Save` can do a forced push, and that can
cause problems if you are using a `post-update` hook to pull into, e.g.,
a website.  You'll get warned about it, but it would still be
inconvenient.  If it happens too often, it would be worthwhile to
replace `git pull` in the hook with `git reset --hard origin/main` or

```
    git pull --no-ff || git reset --hard origin/main
```

It's somewhat moot at this point, because neither of the repos using this
plug-in have working trees on the web host.