Remove `.PHONY: .MM` declaration
`.MM` is a convenient symlink to `$(POST_ARCHIVE)$(MONTHPATH)`.  It was
declared .PHONY, which caused it to be remade every time.  In addition,
because it's a symlink to a directory, `ln -sf` needs to be preceeded by
removing the previous link; the `-f` flag doesn't do the expected thing
in this case.