Add blogging/save-amend-commit for make save to use
Rather than simply replacing the old "Saved on..." message with a new
one, retain all previous messages in the message body, newest first.

The deed is done with the command:

```
GIT_EDITOR=$(TOOLDIR)/blogging/save-amend-commit	\
 NEW_MESSAGE='Saved on $(COMMIT_MSG)' git commit -a --amend;

The simplest thing would be to append the new message to the original

```
echo "$NEW_MESSAGE" >> $1
```

However, git reuses COMMIT_EDITMSG, which typically has a comment block
starting with "\n# Please enter..."  appended to it, and also puts the
history after the original commit message, which might have had more
than one line (e.g. a count of to.do items, in the usual case).

Instead, we _pre_pend the new message, typically `Saved on ...`, and use
a little sed script to remove the blank line after the previous one.
(Preserving the blank line between the _earliest_ message and the
original body text, if any, would require additional sed hackery, which
is left as an exercise for the reader.)

Passing `save-amend-commit` the `--append` option appends the message.
Cleaning up the double spacing can be done with a sed script which,
again, is left as an exercise.