Make templates in blogging not use export
They now get expanded into an auxiliary file, so they won't get picked
up by git if make was invoked with `-n`.

With this arrangement, there is no need for template definitions to be
exported as environment variables.  This may be useful on OSs where the
environment is limited (Windows?)