make report-var et. al. handle quotes
Instead of using `echo` to report values, we use the make function
`$(info)`, which doesn't have a problem with quotes or other shell
special characters.

Since `$(info)` prints as a side effect and expands to the empty string,
we preceed lines that use it with `@:` (where `:` is a shell builtin
that's equivalent to `true`) to keep make from saying "nothing to be
done for...".  Keeping the output clean might be useful for, e.g., using
`make report-var` in a script.