Improve/fix test for changes in save
Make the test a little more clear.  Tried to be clever and use diff, but
that caused more problems than it solved, and wound up needing a pipe
because `git diff -s` has the wrong exit code, and it's not clear that
it would be more efficient.

To cater to the inevitable urge to keep tweaking it, the test has been
pulled out into a recursive variable (i.e. macro) called
`nothing-to-commit`.  It might be worthwhile moving it to defines.make.