add $(COMMIT_OPTS) to first save
This is what `make push` does -- in `Dog` (the to.do log directory) it
appends a line with "undone" item counts to the the commit under
construction.  In `make save`, we only do it if we're _not_ amending a
previous save or push commit, thus keeping save commits less cluttered.

Not that anyone actually looks at them, except on rare occasions.