move .draft.d to .post.d in post:
See previous commit.  The `posted` target is for fixup; the `post`
target does many of the same things, but it's different enough to be
confusing.  This commit applies the same change to `post` that the
previous commit added to `posted`.

At some point it might be a good idea to unify these somehow.