Use -e, not -d, to detect .draft.d
Although `test` dereferences symlinks for _file-oriented_ tests, it
apparently does not consider `-d` to be file-oriented.  (This is
probably implementation-dependent.)  In any case, the `post:` target
needs to use `-e` to detect the presence of .draft.d.

Also, unconditionally remove the previous `.post.d` to keep it from
hanging around, in case `.draft.d` doesn't exist.