Handle a draft resource directory in entry
The resource directory will be $(name).d; it's handled somewhat weirdly
because 1, it might not exist, and 2, if it does it's not going to be
under git control.  Note that things might get weirder still if we
later decide to put the resources under git-annex or something of the
sort.

As things stand now, it's possible that updraft: and redraft: targets
might still have bugs.