Fix `make save` by fixing the commit msg format
It turned out that between September 15th and 16th the default output of
`date` changed.  The previous attempt at fixing this was to reduce the
number of words compared in `make save` to determine whether to amend
the previous commit.  A better fix is to create a make variable called
LONG_DATE, and use that in the default commit message.  Switch from
12-hour to 24-hour time -- that part of the message isn't compared.

If you want to change the date format, you can do a save (for testing)
by setting SAVE_MAY_IGNORE_DATE=1 on the command line.