Improve reporting in check-html and make check
Check for missing files in `blogging/check-html` rather than only in
`entry.make`.  That lets us check .post, and move on to whatever else
the local .config.make wants to check.  We do get a line saying .draft
doesn't exist; that may want to be fixed at some point, but for now it's
better than reporting it as an error.

If .draft doesn't exist, we get an error message, which isn't ideal but
it works.